Added Basic Player Action Handling Logic to Update Game State
For each player action we run apply it to the state and if the action
is valid we return the new state, write it to the serverState and
broadcast the new game state to all subscribers. On the otherhand
if the player action is invalid we simply send an error message
detailing why the game action is invalid back to the originator
of the action.